Create AirTicketPrice.java
Calculates the air ticket price according to the distance and conditions. It receives Distance (KM), age and type of trip (One Way, Round Trip) information from the user. It charges a fee of 0.10 TL / km per distance. First, it calculates the total price of the flight and then applies the following discounts to the customer according to the conditions
If the person is under the age of 12, a 50% discount is applied on the ticket price.
If the person is between the ages of 12 and 24, a 10% discount is applied on the ticket price.Dec.
If the person is older than 65 years, a 30% discount is applied on the ticket price.
If the person has chosen the "Trip Type" round trip, a 20% discount is applied on the ticket price.